% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlastoGram-package.R
\docType{package}
\name{PlastoGram-package}
\alias{PlastoGram-package}
\alias{PlastoGram}
\title{Prediction of subchloroplast localization.}
\description{
PlastoGram is available as R function (\code{\link{predict.plastogram_model}}) or
shiny GUI (\code{\link{PlastoGram_gui}}).
}
\details{
PlastoGram requires the external software HMMER.
}
\references{
...
}
\author{
Maintainer: Katarzyna Sidorczuk <sidorczuk.katarzyna17@gmail.com>
}
\keyword{package}
