% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlastoGram_gui.R
\name{PlastoGram_gui}
\alias{PlastoGram_gui}
\title{PlastoGram Graphical User Interface}
\usage{
PlastoGram_gui()
}
\value{
No return value, called for side effects.
}
\description{
Launches graphical user interface that predicts subchloroplast
localization of proteins.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\seealso{
\code{\link[shiny]{runApp}}
}
