% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmmer.R
\name{read_tblout}
\alias{read_tblout}
\title{Read a tblout HMMER file}
\usage{
read_tblout(file)
}
\arguments{
\item{file}{Filename}
}
\value{
data.frame
}
\description{
This function is a modified version of a read_tblout function 
from the rhmmer package. For the original code see
https://github.com/arendsee/rhmmer/blob/master/R/parse.R
}
