% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plastogram_model}
\alias{predict.plastogram_model}
\title{Predict subchloroplast localization}
\usage{
\method{predict}{plastogram_model}(object, newdata, hmmer_dir, ...)
}
\arguments{
\item{object}{\code{plastogram_model} object.}

\item{newdata}{\code{list} of sequences (for example as given by
\code{\link[biogram]{read_fasta}} or \code{\link{read_txt}}).}

\item{hmmer_dir}{path to the hmmer directory in which \code{hmmsearch} executable 
is located}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object of class \code{plastogram_prediction}, a \code{list} of three
\code{data frame}s containing prediction results:
\describe{
  \item{Lower-order_models_preds}{Prediction results from nine lower-order
  models trained to recognize sequence features associated with specific 
  subchloroplast localization. Data frame with 10 columns and number of rows
  equal to the number of analyzed sequences. The first column contains sequence
  name and the following  For more information on lower-order models
  see Details section.}
  \item{Higher-order_model_preds}{Prediction results from higher-order model
  trained to determine final subchloroplast localization of a given protein
  based on predictions obtained by lower-order models. Data frame with 10 columns
  and number of rows equal to number of analyzed sequences. The first column
  (\code{seq_name}), the following nine columns contain prediction probabilities
  for each of the locations considered by the PlastoGram model. The last column
  (\code{Localization}) contains abbreviation of a predicted location. For more 
  information on higher-order model see Details section.}
  \item{Final_results}{Summary of PlastoGram predictions. Data frame with 3
  columns and number of rows equal to the number of analyzed sequences. The
  columns contain the following information: name of the analyzed sequence,
  predicted localization, probability of the predicted localization (assumes
  values from 0 to 1).}}
}
\description{
Predicts protein subchloroplast localization using the PlastoGram algorithm.
}
\details{
PlastoGram depends on the HMMER software for prediction of signals
responsible for targeting to the thylakoid lumen via Sec and Tat pathways.

PlastoGram lower-order models are responsible for identification of features
characteristic for specific subchloroplast localizations. They include
random forest models based on ngrams (short amino acid motifs):
\describe{
  \item{Nuclear_model}{recognizes nuclear-encoded proteins}
  \item{Membrane_model}{recognizes membrane proteins}
  \item{Nuclear_OM_model}{recognizes nuclear-encoded proteins targeted to 
  plastid outer membrane}
  \item{Nuclear_IM_model}{recognizes nuclear-encoded proteins targeted to 
  plastid inner membrane}
  \item{Nuclear_TM_model}{recognizes nuclear-encoded proteins targeted to 
  plastid thylakoid membrane}
  \item{Plastid_membrane_model}{differentiates plastid-encoded proteins 
  targeted to plastid inner and thylakoid membrane. Prediction values 0 
  indicate thylakoid membrane, whereas 1 inner membrane}
  \item{Nuclear_OM_stroma_model}{differentiates nuclear-encoded proteins 
  targeted to plastid outer membrane and stroma. Prediction values 0 
  indicate stroma, whereas 1 outer membrane}}
and profile HMM models based on HMMER software
\describe{
  \item{Sec_model}{recognizes proteins targeted to the thylakoid lumen
  via Sec pathway}
  \item{Tat_model}{recognizes proteins targeted to the thylakoid lumen
  via Tat pathway}}
}
